+++ packages/build-plugin-component/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.9.2
 
-- [feat] `allowJs` to support compile `.d.ts` for JavaScript who uses jsdoc.
+- [feat] `generateTypesForJs` to support compile `.d.ts` for JavaScript who uses jsdoc.
 - [fix] compile error may occur when compile `.d.ts` for JavaScript. ([#246](https://github.com/ice-lab/iceworks-cli/issues/246))
 
 ## 1.9.1